requirements.txt
asgiref==3.8.1
CacheControl==0.14.2
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.1
Django==5.2
django-cors-headers==4.7.0
djangorestframework==3.16.0
fire==0.7.0
fonttools==4.56.0
google-api-core==2.24.1
google-api-python-client==2.162.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.2
google-cloud-firestore==2.20.1
google-cloud-storage==3.0.0
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.68.0
grpcio==1.70.0
grpcio-status==1.70.0
httplib2==0.22.0
idna==3.10
lxml==5.3.1
msgpack==1.1.0
mysqlclient==2.2.7
numpy==2.2.3
opencv-python-headless==4.11.0.86
packaging==24.2
pdf2docx==0.5.8
pillow==11.1.0
proto-plus==1.26.0
protobuf==5.29.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydub==0.25.1
PyJWT==2.10.1
PyMuPDF==1.25.3
pyparsing==3.2.1
pytesseract==0.3.13
python-docx==1.1.2
requests==2.32.3
rsa==4.9
SpeechRecognition==3.14.1
sqlparse==0.5.3
termcolor==2.5.0
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.3.0
